refactor: move reconcile function update_ranges -> pipeline
